+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 
 Our Keycloak version is working well with PostgreSQL database. For using other SQL databases, text field in database need to be evaluated.
 
-## [Unreleased]
+## [22.0.3-1.24] - 2025-10-09
 
 _keycloakUrl_ realm attribute is needed to be configured for a realm being in OpenId Federation as RP